Use host as VM name in libvirt
This is useful if you use dnsmasq with libvirt/QEMU,
and want to have a fqdn for your virtual machines.
Sadly this is not usable as a terraform resource name,
so we prepend the vm group and the vm name as a unique
identifier for the resource.
Having the vm group as resource name lets the user share
the same pool/storage/cloudinit/etc in multiple vm groups.